python/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memlayer-agent-python"
version = "0.1.0"
description = "Python implementation of MemLayer Agent ADK server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-genai>=1.0.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "structlog>=24.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
